refactor: share common gateway write commands file vs overloaded cli file
