+++ src/cli/commands/gatewayWriteCommands.ts
@@ -13,11 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import prompts from 'prompts';
+
 import { mIOToken } from '../../node/index.js';
 import {
   AddressAndVaultIdCLIWriteOptions,
+  DecreaseDelegateStakeCLIOptions,
   JoinNetworkCLIOptions,
   OperatorStakeCLIOptions,
+  RedelegateStakeCLIOptions,
+  TransferCLIOptions,
   UpdateGatewaySettingsCLIOptions,
   WriteActionCLIOptions,
 } from '../types.js';
@@ -26,10 +31,12 @@ import {
   assertEnoughBalance,
   formatIOWithCommas,
   gatewaySettingsFromOptions,
+  redelegateParamsFromOptions,
   requiredAddressFromOptions,
   requiredMIOFromOptions as requiredMARIOFromOptions,
   requiredStringArrayFromOptions,
   requiredStringFromOptions,
+  requiredTargetAndQuantityFromOptions,
   stringifyJsonForCLIDisplay,
   writeActionTagsFromOptions,
   writeIOFromOptions,
@@ -235,4 +242,119 @@ export async function cancelWithdrawal(o: AddressAndVaultIdCLIWriteOptions) {
     },
     writeActionTagsFromOptions(o),
   );
+}
+
+export async function delegateStake(options: TransferCLIOptions) {
+  const { io, signerAddress } = writeIOFromOptions(options);
+
+  const { target, ioQuantity } = requiredTargetAndQuantityFromOptions(options);
+  const mIOQuantity = ioQuantity.toMIO();
+
+  if (!options.skipConfirmation) {
+    const balance = await io.getBalance({ address: signerAddress });
+
+    if (balance < mIOQuantity.valueOf()) {
+      throw new Error(
+        `Insufficient IO balance for delegating stake. Balance available: ${new mIOToken(balance).toIO()} IO`,
+      );
+    }
+
+    const targetGateway = await io.getGateway({ address: target });
+    if (targetGateway === undefined) {
+      throw new Error(`Gateway not found for address: ${target}`);
+    }
+    if (targetGateway.settings.allowDelegatedStaking === false) {
+      throw new Error(`Gateway does not allow delegated staking: ${target}`);
+    }
+
+    // TODO: could get allow list and assert doesn't exist or user is on it
+
+    // TODO: could read from contract to get current delegated stake if there is none, get contract minimum delegated stake. Then see if the new stake value will satisfy minimum delegated stake for both the target gateway settings min delegate stake and contract min delegated amounts
+
+    const { confirm } = await prompts({
+      type: 'confirm',
+      name: 'confirm',
+      message: `Target Gateway:\n${JSON.stringify(targetGateway, null, 2)}\n\nAre you sure you want to delegate ${formatIOWithCommas(ioQuantity)} IO to ${target}?`,
+    });
+
+    if (!confirm) {
+      return { message: 'Delegate stake aborted by user' };
+    }
+  }
+
+  const result = await io.delegateStake(
+    {
+      target,
+      stakeQty: ioQuantity.toMIO(),
+    },
+    writeActionTagsFromOptions(options),
+  );
+
+  const output = {
+    senderAddress: signerAddress,
+    transferResult: result,
+    message: `Successfully delegated ${formatIOWithCommas(ioQuantity)} IO to ${target}`,
+  };
+
+  return output;
+}
+
+export async function decreaseDelegateStake(
+  options: DecreaseDelegateStakeCLIOptions,
+) {
+  const io = writeIOFromOptions(options).io;
+  const { target, ioQuantity } = requiredTargetAndQuantityFromOptions(options);
+  const instant = options.instant ?? false;
+
+  // TODO: Could assert sender is a delegate with enough stake to decrease
+  // TODO: Could assert new target stake meets contract and target gateway minimums
+  // TODO: Could present confirmation prompt with any fee for instant withdrawal (50% of the stake is put back into protocol??)
+
+  await assertConfirmationPrompt(
+    `Are you sure you'd like to decrease delegated stake of ${formatIOWithCommas(ioQuantity)} IO on gateway ${target}?`,
+    options,
+  );
+
+  const result = await io.decreaseDelegateStake({
+    target,
+    decreaseQty: ioQuantity.toMIO(),
+    instant,
+  });
+
+  const output = {
+    targetGateway: target,
+    decreaseDelegateStakeResult: result,
+    message: `Successfully decreased delegated stake of ${formatIOWithCommas(
+      ioQuantity,
+    )} IO to ${target}`,
+  };
+
+  return output;
+}
+
+export async function redelegateStake(options: RedelegateStakeCLIOptions) {
+  const io = writeIOFromOptions(options).io;
+  const params = redelegateParamsFromOptions(options);
+
+  // TODO: Could assert target gateway exists
+  // TODO: Could do assertion on source has enough stake to redelegate
+  // TODO: Could do assertions on source/target min delegate stakes are met
+
+  await assertConfirmationPrompt(
+    `Are you sure you'd like to redelegate stake of ${formatIOWithCommas(params.stakeQty.toIO())} IO from ${params.source} to ${params.target}?`,
+    options,
+  );
+
+  const result = await io.redelegateStake(params);
+
+  const output = {
+    sourceGateway: params.source,
+    targetGateway: params.target,
+    redelegateStakeResult: result,
+    message: `Successfully re-delegated stake of ${formatIOWithCommas(
+      params.stakeQty.toIO(),
+    )} IO from ${params.source} to ${params.target}`,
+  };
+
+  return output;
 }
